Refinado codigo js para apresentar info das turmas, ajustado logica das rotas da api que retorna todas as turmas e que retornam grupos e alunos de um turma_id
